Move sinusoidal movement to test_trafo_manager
