+++ pytransform3d/transform_manager/_temporal_transform_manager.py
@@ -133,7 +133,6 @@ class PandasTimeseriesTransform(TimeVaryingTransform):
         return pq
 
 
-
 class TemporalTransformManager(TransformGraphBase):
     """Manage time-varying transformations.
